\name{parquet_read}
\alias{parquet_read}

\title{
   Read a Parquet file into a data.frame
}
\description{
  Converts the contents of the named Parquet file to a R data frame.
}
\usage{
parquet_read(file) 
}
\arguments{
  \item{file}{Path to a Parquet file.}

 
 }
\value{
  A \code{data.frame} with the file's contents.
}
